--- a/src/mlpack/core/optimizers/sdp/lrsdp_function.hpp
+++ b/src/mlpack/core/optimizers/sdp/lrsdp_function.hpp
@@ -22,6 +22,15 @@ namespace optimization {
 
 /**
  * The objective function that LRSDP is trying to optimize.
+ *
+ * Note: LRSDPfunction is designed and implemented to specifically work
+ * with a combination of AugLagrangian and L-BFGS optimizer. Implemenation
+ * of LRSDPFunction includes caching R * R^T matrix in order to avoid redundant
+ * computations(Specifically with above two optimizers) of R * R^T matrix
+ * through AugLagrangian optimizer call, as only L-BFGS takes part in updating
+ * R(coordinates) matrix. So, be careful while using LRSDP with some other
+ * optimizer. You may need to modify caching process of R * R^T matrix.
+ * See EvaluateImpl() in lrsdp_function_impl.hpp for more details.
  */
 template <typename SDPType>
 class LRSDPFunction