Reduces number of computations by taking `R^T * C ` first in objective function and `R^T * A` in constraints converting objective function to `Tr((R^T * C) * R)` and similarly constraints to `Tr((R^T * A) * R)`.