@rcurtin 2018-01-29
If I am understanding right, this optimizations hinges around the fact that for (A*R^T*R), the matrix (A*R^T) is significantly smaller than (R^T*R) and therefore it's much quicker to compute (A*R^T)*R than A*(R^T*R).  This would happen when the number of constraints is fewer than the number of parameters, which I think is generally true.  Let me know if I have understood that right.  (I know I have written the algebra in the original form not the column-major Armadillo form where everything is transposed, I hope that is not confusing.)

Do you have any quick runtime tests to show that the evaluations and gradient calculations are now faster?  Simply running one of the test cases on LRSDP and comparing both versions over a handful of runs may be sufficient.