@manish7294 2018-01-29
@rcurtin here are the running times for multiple run of Johnson844LovaszThetaSDP test.

> i=1; while(true); do echo -n "Run" $i":  "; bin/mlpack_test -t LRSDPTest/Johnson844LovaszThetaSDP 2>&1 | grep -v 'Running ' | grep -v '^[ ]*$' | grep -v 'No errors'; i=$(($i + 1)); sleep 1; done

Previous:
```
Run 1:  2.76081s
Run 2:  2.81437s
Run 3:  2.85808s
Run 4:  2.79318s
Run 5:  2.83648s
Run 6:  2.82382s
Run 7:  2.82247s
Run 8:  2.78188s
Run 9:  2.80789s
Run 10:  2.91141s
Run 11:  2.8192s
Run 12:  2.85181s
Run 13:  2.78982s
Run 14:  2.83226s
Run 15:  2.84832s
Run 16:  2.7895s
Run 17:  2.77255s
Run 18:  2.86243s
Run 19:  2.8154s
Run 20:  2.90054s
```

Proposed:                                 
```
Run 1:  1.32478s
Run 2:  1.31208s
Run 3:  1.32616s
Run 4:  1.31512s
Run 5:  1.31336s
Run 6:  1.30926s
Run 7:  1.31946s
Run 8:  1.32875s
Run 9:  1.31783s
Run 10:  1.32423s
Run 11:  1.31942s
Run 12:  1.33508s
Run 13:  1.31668s
Run 14:  1.3287s
Run 15:  1.35265s
Run 16:  1.31307s
Run 17:  1.3154s
Run 18:  1.32277s
Run 19:  1.31655s
Run 20:  1.29777s

```