@manish7294 2018-01-30
_Running Time for different versions over multilple runs:_

**# GaussianMatrixSensingSDP**

Run Command Used:

> i=1; while(true); do echo -n "Run" $i": "; bin/mlpack_test -t LRSDPTest/GaussianMatrixSensingSDP 2>&1 | grep -v 'Running ' | grep -v '^[ ]*$' | grep -v 'No errors'; i=$(($i + 1)); sleep 1; done

Previous:
```
Run 1: 12.989s   (GaussianMatrixSensingSDP)
Run 2: 12.9791s   (GaussianMatrixSensingSDP)
Run 3: 12.9452s   (GaussianMatrixSensingSDP)
Run 4: 12.8985s   (GaussianMatrixSensingSDP)
Run 5: 13.0074s   (GaussianMatrixSensingSDP)
Run 6: 12.9212s   (GaussianMatrixSensingSDP)
Run 7: 13.0073s   (GaussianMatrixSensingSDP)
Run 8: 12.9725s   (GaussianMatrixSensingSDP)
Run 9: 12.899s   (GaussianMatrixSensingSDP)
Run 10: 12.9422s   (GaussianMatrixSensingSDP)

```

Using optimization only on dense matrices:

```
 const arma::mat rrt = coordinates * trans(coordinates);
  if (index < SDP().NumSparseConstraints())
    return accu(SDP().SparseA()[index] % rrt) - SDP().SparseB()[index];
  const size_t index1 = index - SDP().NumSparseConstraints();
  return trace((trans(coordinates) * SDP().DenseA()[index1]) * coordinates)
                 - SDP().DenseB()[index1];
```

```
Run 1: 12.4094s   (GaussianMatrixSensingSDP)
Run 2: 12.8643s   (GaussianMatrixSensingSDP)
Run 3: 12.1417s   (GaussianMatrixSensingSDP)
Run 4: 12.0798s   (GaussianMatrixSensingSDP)
Run 5: 12.6487s   (GaussianMatrixSensingSDP)
Run 6: 12.0841s   (GaussianMatrixSensingSDP)
Run 7: 11.9463s   (GaussianMatrixSensingSDP)
Run 8: 12.2213s   (GaussianMatrixSensingSDP)
Run 9: 12.3982s   (GaussianMatrixSensingSDP)
Run 10: 11.9218s   (GaussianMatrixSensingSDP)
```

Optimized ais[i]  (Here C is also optimized) _Tried only 5 times_ :

```
Run 1: 178.129s   (GaussianMatrixSensingSDP)
Run 2: 179.623s   (GaussianMatrixSensingSDP)
Run 3: 180.395s   (GaussianMatrixSensingSDP)
Run 4: 178.663s   (GaussianMatrixSensingSDP)
Run 5: 183.692s   (GaussianMatrixSensingSDP)
```

Current PR version:

```
Run 1: 11.6893s   (GaussianMatrixSensingSDP)
Run 2: 11.7374s   (GaussianMatrixSensingSDP)
Run 3: 11.7082s   (GaussianMatrixSensingSDP)
Run 4: 11.6893s   (GaussianMatrixSensingSDP)
Run 5: 11.6591s   (GaussianMatrixSensingSDP)
Run 6: 11.6397s   (GaussianMatrixSensingSDP)
Run 7: 11.6419s   (GaussianMatrixSensingSDP)
Run 8: 11.597s   (GaussianMatrixSensingSDP)
Run 9: 11.6155s   (GaussianMatrixSensingSDP)
Run 10: 11.6926s   (GaussianMatrixSensingSDP)
```




**# Johnson844LovaszThetaSDP**




Run Command Used:

> i=1; while(true); do echo -n "Run" $i": "; bin/mlpack_test -t LRSDPTest/Johnson844LovaszThetaSDP LRSDP/GaussianMatrixSensingSDP 2>&1 | grep -v 'Running ' | grep -v '^[ ]*$' | grep -v 'No errors'; i=$(($i + 1)); sleep 1; done

Previous:

```
Run 1: 2.90536s   (Johnson844LovaszThetaSDP)
Run 2: 2.93401s   (Johnson844LovaszThetaSDP)
Run 3: 3.15133s   (Johnson844LovaszThetaSDP)
Run 4: 2.87859s   (Johnson844LovaszThetaSDP)
Run 5: 2.9493s   (Johnson844LovaszThetaSDP)
Run 6: 2.98959s   (Johnson844LovaszThetaSDP)
Run 7: 2.96007s   (Johnson844LovaszThetaSDP)
Run 8: 2.8916s   (Johnson844LovaszThetaSDP)
Run 9: 2.90623s   (Johnson844LovaszThetaSDP)
Run 10: 2.88496s   (Johnson844LovaszThetaSDP)
```

Optimized ais[i]  (Here C is also optimized):

```
Run 1: 2.4925s   (Johnson844LovaszThetaSDP)
Run 2: 2.48819s   (Johnson844LovaszThetaSDP)
Run 3: 2.49341s   (Johnson844LovaszThetaSDP)
Run 4: 2.49682s   (Johnson844LovaszThetaSDP)
Run 5: 2.50309s   (Johnson844LovaszThetaSDP)
Run 6: 2.50949s   (Johnson844LovaszThetaSDP)
Run 7: 2.58773s   (Johnson844LovaszThetaSDP)
Run 8: 2.54184s   (Johnson844LovaszThetaSDP)
Run 9: 2.51936s   (Johnson844LovaszThetaSDP)
Run 10: 2.49226s   (Johnson844LovaszThetaSDP)
```

Using optimization only on dense matrices:

```
 const arma::mat rrt = coordinates * trans(coordinates);
  if (index < SDP().NumSparseConstraints())
    return accu(SDP().SparseA()[index] % rrt) - SDP().SparseB()[index];
  const size_t index1 = index - SDP().NumSparseConstraints();
  return trace((trans(coordinates) * SDP().DenseA()[index1]) * coordinates)
                 - SDP().DenseB()[index1];
```

```
Run 1: 2.03692s   (Johnson844LovaszThetaSDP)
Run 2: 2.05736s   (Johnson844LovaszThetaSDP)
Run 3: 2.08177s   (Johnson844LovaszThetaSDP)
Run 4: 2.03506s   (Johnson844LovaszThetaSDP)
Run 5: 2.09492s   (Johnson844LovaszThetaSDP)
Run 6: 2.08166s   (Johnson844LovaszThetaSDP)
Run 7: 2.05037s   (Johnson844LovaszThetaSDP)
Run 8: 2.04741s   (Johnson844LovaszThetaSDP)
Run 9: 2.0342s   (Johnson844LovaszThetaSDP)
Run 10: 2.0469s   (Johnson844LovaszThetaSDP)

```


Current PR version :

```
Run 1: 1.32836s   (Johnson844LovaszThetaSDP)
Run 2: 1.59912s   (Johnson844LovaszThetaSDP)
Run 3: 1.34011s   (Johnson844LovaszThetaSDP)
Run 4: 1.33653s   (Johnson844LovaszThetaSDP)
Run 5: 1.37482s   (Johnson844LovaszThetaSDP)
Run 6: 1.42944s   (Johnson844LovaszThetaSDP)
Run 7: 1.37366s   (Johnson844LovaszThetaSDP)
Run 8: 1.37739s   (Johnson844LovaszThetaSDP)
Run 9: 1.37261s   (Johnson844LovaszThetaSDP)
Run 10: 1.34749s   (Johnson844LovaszThetaSDP)
```

My Inference: The optimization is true for both sparse and dense matrices.

@rcurtin Here is what I found. Please suggest if anything else can be done or I if am wrong somewhere.
This comment might be a bit lengthy one :)